+++ bson/src/main/scala/org/mongodb/scala/bson/annotations/BsonProperty.scala
@@ -1,11 +1,11 @@
 /*
- * Copyright 2015 MongoDB, Inc.
+ * Copyright 2017 MongoDB, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,6 +17,10 @@
 package org.mongodb.scala.bson.annotations
 
 import scala.annotation.StaticAnnotation
-import scala.language.experimental.macros
 
-class Key(val key: String) extends StaticAnnotation
+/**
+ * Annotation to change the stored key of a property
+ *
+ * @param key the key for the stored property
+ */
+case class BsonProperty(key: String) extends StaticAnnotation