+++ docs/reference/content/changelog.md
@@ -11,7 +11,8 @@ title = "Changelog"
 Changes between released versions
 
 ### 2.2.0
-  
+
+  * Added BsonProperty annotation [SCALA-321](https://jira.mongodb.org/browse/SCALA-321)
   * Updated Mongodb Driver Async dependency to [3.5.0](https://jira.mongodb.org/browse/SCALA-335)
   * CaseClassCodec - Added support for internal vals. [SCALA-314](https://jira.mongodb.org/browse/SCALA-314)
   * CaseClassCodec - Added support for default values. [SCALA-313](https://jira.mongodb.org/browse/SCALA-313)